fix(log): update file permissions before removal
- Add chmod command to set file permissions to 0755 before removing
- This change ensures proper file permissions are set before deletion

+++ zlog/log_config.go
@@ -133,6 +133,11 @@ func (l *LogsConfig) PruneLogs() (string, error) {
 		return "", nil
 	}
 
+	errChangeChmod := os.Chmod(localPath, 0755)
+	if errChangeChmod != nil {
+		return localPath, errChangeChmod
+	}
+
 	errRemoveAll := os.RemoveAll(localPath)
 	if errRemoveAll != nil {
 		return localPath, errRemoveAll
